Feature Update - Added 0.5 rating system

+++ src/main/java/com/atmymelo/atmymelobackend/dto/AlbumDTOs/FavoriteDTO/FavoriteGenreResponseDTO.java
@@ -4,7 +4,7 @@ public record FavoriteGenreResponseDTO(
         String albumId,
         String title,
         String artist,
-        int rating,
+        Double rating,
         Integer releaseDate,
         String imageUrl
 ) {}
