+++ src/main/java/com/atmymelo/atmymelobackend/dto/AlbumDTOs/RatingDTO/RatingRequestDTO.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.Size;
 
 public record RatingRequestDTO(
         @Size(min = 1, max = 5)
-        int rating
+        Double rating
 ) {
 }
