+++ src/main/java/com/atmymelo/atmymelobackend/dto/TrackDTOs/UserTrackResponseDto.java
@@ -1,7 +1,7 @@
 package com.atmymelo.atmymelobackend.dto.TrackDTOs;
 
 public record UserTrackResponseDto(
-        Integer userRating,
+        Double userRating,
         Double avgRating
 ) {
 }
